+++ src/Controller/AdministrateurController.php
@@ -10,14 +10,24 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\PasswordHasher\Hasher\UserPasswordHasherInterface;
+
 
 #[Route('/administrateur')]
 class AdministrateurController extends AbstractController
 {
+    public function __construct(UserPasswordHasherInterface $passwordHasher)
+    {
+        $this->passwordHasher = $passwordHasher;
+    }
     #[Route('/', name: 'app_administrateur_index', methods: ['GET'])]
     public function index(EntityManagerInterface $entityManager): Response
     {
         $usr = $this->getUser();
+        if(!isset($usr->getRoles()["app_administrateur_index"]) && !isset($usr->getRoles()['ROLE_MASTER'])) {
+            $this->user = $usr;
+            return $this->redirectToRoute('app_club_show', ["id" => $usr->getClubid()->getId()], Response::HTTP_SEE_OTHER);
+        }
         $administrateurs = $entityManager
             ->getRepository(Administrateur::class)
             ->findAll();
@@ -35,6 +45,10 @@ class AdministrateurController extends AbstractController
     public function new(Request $request, EntityManagerInterface $entityManager): Response
     {
         $usr = $this->getUser();
+        if(!isset($usr->getRoles()["app_administrateur_new"]) && !isset($usr->getRoles()['ROLE_MASTER'])) {
+            $this->user = $usr;
+            return $this->redirectToRoute('app_club_show', ["id" => $usr->getClubid()->getId()], Response::HTTP_SEE_OTHER);
+        }
         $administrateur = new Administrateur();
         $form = $this->createForm(AdministrateurType::class, $administrateur);
         $form->handleRequest($request);
@@ -46,6 +60,10 @@ class AdministrateurController extends AbstractController
             $administrateur->setCreatedAt(new \DateTime());
             $administrateur->setUpdatedAt(new \DateTime());
             $administrateur->setRoles(['ROLE_ADMIN']);
+            $administrateur->setPassword($this->passwordHasher->hashPassword(
+                $administrateur,
+                $form->get('password')->getData()
+            ));
             $entityManager->persist($administrateur);
             $entityManager->flush();
 
@@ -79,6 +97,10 @@ class AdministrateurController extends AbstractController
     public function edit(Request $request, Administrateur $administrateur, EntityManagerInterface $entityManager): Response
     {
         $usr = $this->getUser();
+        if(!isset($usr->getRoles()["app_administrateur_edit"]) && !isset($usr->getRoles()['ROLE_MASTER'])) {
+            $this->user = $usr;
+            return $this->redirectToRoute('app_club_show', ["id" => $usr->getClubid()->getId()], Response::HTTP_SEE_OTHER);
+        }
         $form = $this->createForm(AdministrateurType::class, $administrateur);
         $form->handleRequest($request);
         $sections = $entityManager
@@ -86,6 +108,10 @@ class AdministrateurController extends AbstractController
             ->findBy(['clubid' => $this->getUser()->getClubid()]);
 
         if ($form->isSubmitted() && $form->isValid()) {
+            $administrateur->setPassword($this->passwordHasher->hashPassword(
+                $administrateur,
+                $form->get('password')->getData()
+            ));
             $entityManager->flush();
 
             $this->user = $usr;
@@ -104,6 +130,10 @@ class AdministrateurController extends AbstractController
     public function delete(Request $request, Administrateur $administrateur, EntityManagerInterface $entityManager): Response
     {
         $usr = $this->getUser();
+        if(!isset($usr->getRoles()["app_administrateur_delete"]) && !isset($usr->getRoles()['ROLE_MASTER'])) {
+            $this->user = $usr;
+            return $this->redirectToRoute('app_club_show', ["id" => $usr->getClubid()->getId()], Response::HTTP_SEE_OTHER);
+        }
         if ($this->isCsrfTokenValid('delete'.$administrateur->getId(), $request->request->get('_token'))) {
             $entityManager->remove($administrateur);
             $entityManager->flush();
