+++ src/Controller/SupervisorController.php
@@ -10,18 +10,28 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\PasswordHasher\Hasher\UserPasswordHasherInterface;
+
 
 #[Route('/supervisor')]
 class SupervisorController extends AbstractController
 {
+    public function __construct(UserPasswordHasherInterface $passwordHasher)
+    {
+        $this->passwordHasher = $passwordHasher;
+    }
     #[Route('/', name: 'app_supervisor_index', methods: ['GET'])]
     public function index(EntityManagerInterface $entityManager): Response
     {
         $usr = $this->getUser();
+        if(!isset($usr->getRoles()["app_supervisor_index"]) && !isset($usr->getRoles()['ROLE_MASTER'])) {
+            $this->user = $usr;
+            return $this->redirectToRoute('app_club_show', ["id" => $usr->getClubid()->getId()], Response::HTTP_SEE_OTHER);
+        }
         $supervisors = $entityManager
             ->getRepository(Supervisor::class)
             ->findAll();
-            $sections = $entityManager
+        $sections = $entityManager
             ->getRepository(Section::class)
             ->findBy(['clubid' => $this->getUser()->getClubid()]);
 
@@ -36,6 +46,10 @@ class SupervisorController extends AbstractController
     public function new(Request $request, EntityManagerInterface $entityManager): Response
     {
         $usr = $this->getUser();
+        if(!isset($usr->getRoles()["app_supervisor_new"]) && !isset($usr->getRoles()['ROLE_MASTER'])) {
+            $this->user = $usr;
+            return $this->redirectToRoute('app_club_show', ["id" => $usr->getClubid()->getId()], Response::HTTP_SEE_OTHER);
+        }
         $supervisor = new Supervisor();
         $form = $this->createForm(SupervisorType::class, $supervisor);
         $form->handleRequest($request);
@@ -44,7 +58,19 @@ class SupervisorController extends AbstractController
             ->findBy(['clubid' => $this->getUser()->getClubid()]);
 
         if ($form->isSubmitted() && $form->isValid()) {
+
+            $supervisor->setCreatedAt(new \DateTime());
+            $supervisor->setUpdatedAt(new \DateTime());
             $supervisor->setRoles(['ROLE_SUPERVISOR']);
+            if(!empty($form->get('clubid')->getData())) {
+                $supervisor->setClubid($form->get('clubid')->getData());
+            } else {
+                $supervisor->setClubid($usr->getClubid());
+            }
+            $supervisor->setPassword($this->passwordHasher->hashPassword(
+                $supervisor,
+                $form->get('password')->getData()
+            ));
             $entityManager->persist($supervisor);
             $entityManager->flush();
 
@@ -78,6 +104,10 @@ class SupervisorController extends AbstractController
     public function edit(Request $request, Supervisor $supervisor, EntityManagerInterface $entityManager): Response
     {
         $usr = $this->getUser();
+        if(!isset($usr->getRoles()["app_supervisor_edit"]) && !isset($usr->getRoles()['ROLE_MASTER'])) {
+            $this->user = $usr;
+            return $this->redirectToRoute('app_club_show', ["id" => $usr->getClubid()->getId()], Response::HTTP_SEE_OTHER);
+        }
         $form = $this->createForm(SupervisorType::class, $supervisor);
         $form->handleRequest($request);
         $sections = $entityManager
@@ -85,6 +115,10 @@ class SupervisorController extends AbstractController
             ->findBy(['clubid' => $this->getUser()->getClubid()]);
 
         if ($form->isSubmitted() && $form->isValid()) {
+            $supervisor->setPassword($this->passwordHasher->hashPassword(
+                $supervisor,
+                $form->get('password')->getData()
+            ));
             $entityManager->flush();
 
             $this->user = $usr;
@@ -103,6 +137,10 @@ class SupervisorController extends AbstractController
     public function delete(Request $request, Supervisor $supervisor, EntityManagerInterface $entityManager): Response
     {
         $usr = $this->getUser();
+        if(!isset($usr->getRoles()["app_supervisor_delete"]) && !isset($usr->getRoles()['ROLE_MASTER'])) {
+            $this->user = $usr;
+            return $this->redirectToRoute('app_club_show', ["id" => $usr->getClubid()->getId()], Response::HTTP_SEE_OTHER);
+        }
         if ($this->isCsrfTokenValid('delete'.$supervisor->getId(), $request->request->get('_token'))) {
             $entityManager->remove($supervisor);
             $entityManager->flush();
